Added a few tests for session in the SDK
